gcc 2.95 fix. Massage the ints to the same sizes. CCMAIL: koffice-devel@kde.org
svn path=/trunk/koffice/; revision=430211

+++ krita/plugins/cimg/CImg.h
@@ -5135,7 +5135,7 @@ namespace cimg_library {
       if (ny1<0 || ny0>=dimy()) return *this;
       if (ny0<0) { nx0-=ny0*(nx1-nx0)/(ny1-ny0); ny0=0; }
       if (ny1>=dimy()) { nx1+=(ny1-dimy())*(nx0-nx1)/(ny1-ny0); ny1=dimy()-1;}
-      const unsigned int dmax = (unsigned int)cimg::max(std::abs(nx1-nx0),ny1-ny0), whz = width*height*depth;
+      const unsigned int dmax = (unsigned int)cimg::max(std::abs((long int)(nx1-nx0)),(long int)(ny1-ny0)), whz = width*height*depth;
       const float px = dmax?(nx1-nx0)/(float)dmax:0, py = dmax?(ny1-ny0)/(float)dmax:0;
       float x = (float)nx0, y = (float)ny0;
       if (opacity>=1) for (unsigned int t=0; t<=dmax; t++) {
